simple Qt and LLVM instead of provisioner installations vga variable support pure offline installation cache extra downloads fix virtualbox postinst hang

+++ scripts/kernel.sh
@@ -5,13 +5,31 @@ export DEBIAN_FRONTEND=noninteractive
 
 test "$i386" = true || dpkg --remove-architecture i386
 
+sed -i -e 's/GRUB_CMDLINE_LINUX_DEFAULT="quiet splash"/GRUB_CMDLINE_LINUX_DEFAULT=""/' -e "s/GRUB_CMDLINE_LINUX=\"net.ifnames=0 nousb noplymouth\"/GRUB_CMDLINE_LINUX=\"vga=$vga net.ifnames=0 nousb noplymouth\"/"  -e 's/GRUB_TIMEOUT=10/GRUB_TIMEOUT=0/' /etc/default/grub
+# high res text
+sed -i -e 's/FONTFACE="VGA"/FONTFACE="TerminusBold"/' -e 's/FONTSIZE="8x16"/FONTSIZE="8x14"/' /etc/default/console-setup
+
+rm -rf /lib/modules/*-generic/kernel/ubuntu/vbox
+
+# do not clear the boot console
+mkdir /etc/systemd/system/getty@tty1.service.d
+echo -e "[Service]\nTTYVTDisallocate=no" > /etc/systemd/system/getty@tty1.service.d/noclear.conf
+
+# disable rev lookups
+echo "UseDNS no" >> /etc/ssh/sshd_config
+
+# move cached packages to destination
+mv /tmp/aptcache/* /var/cache/apt/archives
+
 # select mirror
 if test "$mirror" = auto; then
   mirror="mirror://mirrors.ubuntu.com/mirrors.txt"
-  if test -z "$http_proxy"; then
-    wget http://ftp.us.debian.org/debian/pool/main/n/netselect/netselect_0.3.ds1-28_amd64.deb # netselect-apt_0.3.ds1-28_all.deb
-    dpkg -i netselect_0.3.ds1-28_amd64.deb
-    rm netselect_0.3.ds1-28_amd64.deb
+  if test -z "$http_proxy" -a "$offline" = false; then
+    netselect=/var/cache/apt/archives/netselect_0.3.ds1-28_amd64.deb
+    if test ! -f $netselect; then
+      wget -O $netselect http://ftp.us.debian.org/debian/pool/main/n/netselect/netselect_0.3.ds1-28_amd64.deb # netselect-apt_0.3.ds1-28_all.deb
+    fi
+    dpkg -i $netselect
     mirrors=`wget -q -O- https://launchpad.net/ubuntu/+archivemirrors | grep -P -B8 "statusUP|statusSIX" |  grep -o -P "(f|ht)tp.*\"" | tr '"\n' '  '`
     fmirror=`netselect -s1 -t20 $mirrors 2>/dev/null | awk '{print $2;}'`
     test -n "$fmirror" && mirror=$fmirror
@@ -35,35 +53,21 @@ deb $mirror xenial-backports main restricted universe multiverse
 # deb-src http://archive.canonical.com/ubuntu xenial partner
 EOF
 
-# update apt sources
-apt-get update
+if test "$offline" = false; then
+  # update apt sources
+  apt-get update
 
-# move cached packages to destination
-mv /tmp/aptcache/* /var/cache/apt/archives
+  # Update to the latest kernel
+  apt-get install -y linux-generic linux-image-generic #zerofree
 
-sed -i -e 's/GRUB_CMDLINE_LINUX_DEFAULT="quiet splash"/GRUB_CMDLINE_LINUX_DEFAULT=""/' -e 's/GRUB_TIMEOUT=10/GRUB_TIMEOUT=0/' /etc/default/grub
-# high res text
-sed -i -e 's/FONTFACE="VGA"/FONTFACE="TerminusBold"/' -e 's/FONTSIZE="8x16"/FONTSIZE="8x14"/' /etc/default/console-setup
+  test "$kupgrade" = true || deb_linux_generic="linux-headers-generic linux-image-generic"
 
-# Update to the latest kernel
-apt-get install -y linux-generic linux-image-generic #zerofree
+  apt-get purge -y plymouth-theme-ubuntu-text \
+   linux-image-4.4.0-31-generic linux-image-extra-4.4.0-31-generic \
+   linux-headers-4.4.0-31 linux-headers-4.4.0-31-generic $deb_linux_generic
 
-rm -rf /lib/modules/*-generic/kernel/ubuntu/vbox
-
-test "$kupgrade" = true || deb_linux_generic="linux-headers-generic linux-image-generic"
-
-apt-get purge -y plymouth-theme-ubuntu-text \
- linux-image-4.4.0-31-generic linux-image-extra-4.4.0-31-generic \
- linux-headers-4.4.0-31 linux-headers-4.4.0-31-generic $deb_linux_generic
-
-test "$x11" = true && apt-get install -y xserver-xorg-core # required for virtualbox x11 install
-
-# do not clear the boot console
-mkdir /etc/systemd/system/getty@tty1.service.d
-echo -e "[Service]\nTTYVTDisallocate=no" > /etc/systemd/system/getty@tty1.service.d/noclear.conf
-
-# disable rev lookups
-echo "UseDNS no" >> /etc/ssh/sshd_config
+  test "$x11" = true && apt-get install -y xserver-xorg-core # required for virtualbox x11 install
+fi
 
 # Reboot with the new kernel
 shutdown -r now
@@ -71,4 +75,3 @@ sleep 5
 sync
 reboot -f # in case it hangs
 exit 0
-
